Task/requirements.txt
annotated-types
anyio
async-timeout
asyncpg
bcrypt
boto3
botocore
certifi
cffi
charset-normalizer
click
cryptography
dnspython
ecdsa
email_validator
exceptiongroup
fastapi
ffmpeg
greenlet
h11
httpcore
httpx
idna
Jinja2
jmespath
jose
jwt
localstack
localstack-client
localstack-core
localstack-ext
MarkupSafe
mutagen
moto
passlib
pillow
pyasn1
pycparser
pydantic
pydantic-settings
pydantic_core
pydub
PyJWT
python-dateutil
python-dotenv
python-jose
python-multipart
PyYAML
requests
responses
rsa
s3transfer
six
sniffio
SQLAlchemy
starlette
typing-inspection
typing_extensions
urllib3
uvicorn
Werkzeug
xmltodict
